requirements.txt
fastapi
uvicorn
openai
supabase
python-dotenv
pydantic
pydantic-settings
